+++ src/bootstrap.jl
@@ -161,7 +161,7 @@ end
 
 """
     parametricbootstrap([rng::AbstractRNG], nsamp::Integer, m::MixedModel{T}, ftype=T;
-        β = coef(m), σ = m.σ, θ = m.θ, hide_progress=false, optsum_overrides=(;))
+        β = coef(m), σ = m.σ, θ = m.θ, progress=true, optsum_overrides=(;))
 
 Perform `nsamp` parametric bootstrap replication fits of `m`, returning a `MixedModelBootstrap`.
 
@@ -194,9 +194,16 @@ function parametricbootstrap(
     σ=morig.σ,
     θ::AbstractVector=morig.θ,
     use_threads::Bool=false,
-    hide_progress::Bool=false,
+    progress::Bool=true,
+    hide_progress::Union{Bool,Nothing}=nothing,
     optsum_overrides=(;),
 ) where {T}
+    if !isnothing(hide_progress)
+        Base.depwarn("`hide_progress` is deprecated, please use `progress` instead." *
+                     "NB: `progress` is a positive action, i.e. `progress=true` means show the progress bar.",
+                     :parametricbootstrap; force=true)
+        progress = !hide_progress
+    end
     if σ !== missing
         σ = T(σ)
     end
@@ -216,7 +223,7 @@ function parametricbootstrap(
         "use_threads is deprecated and will be removed in a future release",
         :parametricbootstrap,
     )
-    samp = replicate(n; hide_progress=hide_progress) do
+    samp = replicate(n; progress) do
         simulate!(rng, m; β, σ, θ)
         refit!(m; progress=false)
         # @info "" m.optsum.feval